requirements.txt
Adafruit-DHT=1.4.0
paho-mqtt==1.4.0
sdnotify==0.3.1
colorama==0.4.1
Unidecode==1.0.23
